template changes and first partial styling draft
this is the basic styling for viewing an entry. styling for other
pages (such as editing an entry) is not yet done.

there are some things in here that jdpage doesn't agree with from
a code philosophy standpoint. i'm leaving them in until i have a
full working prototype and then i'll be considering compromises at
a later date.

changes of note:

 * importing jason pamental's font scale.
 * replace normalize with reset, since all font sizes are explicitly
   defined now.
 * templates move the body tag out of main and into sub-templates,
   so that classes can be applied.
 * mathjax is now only loaded on the entry page, and loaded at the
   end of the document to speed up download times.
 * pages which have a nav now have the same nav duplicated in the
   header.